requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.9
langchain-community==0.0.27
langchain-google-genai==0.0.11
faiss-cpu==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
jinja2==3.1.3